+++ CHANGELOG.md
@@ -23,6 +23,7 @@ https://github.com/nwnxee/unified/compare/build8193.37.13...HEAD
 - Creature: GetNumberOfBonusSpells(), ModifyNumberBonusSpells()
 - Store: GetBlackMarket(), SetBlackMarket()
 - Util: SetStartingLocation()
+- Object: GetLocalizedDescription(), SetLocalizedDescription()
 
 ### Changed
 - Damage: Added bRangedAttack to the NWNX_Damage_AttackEventData struct.